made google maps render but by reverting to making a static api request from within App.jsx
